+++ familyhelper-ui-view/packages/component-ckeditor/src/ckeditor.js
@@ -1,0 +1,158 @@
+/**
+ * @license Copyright (c) 2014-2022, CKSource - Frederico Knabben. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
+ */
+import ClassicEditor from '@ckeditor/ckeditor5-editor-classic/src/classiceditor.js';
+import Alignment from '@ckeditor/ckeditor5-alignment/src/alignment.js';
+import Autoformat from '@ckeditor/ckeditor5-autoformat/src/autoformat.js';
+import BlockQuote from '@ckeditor/ckeditor5-block-quote/src/blockquote.js';
+import Bold from '@ckeditor/ckeditor5-basic-styles/src/bold.js';
+import CKFinderUploadAdapter from '@ckeditor/ckeditor5-adapter-ckfinder/src/uploadadapter.js';
+import Code from '@ckeditor/ckeditor5-basic-styles/src/code.js';
+import CodeBlock from '@ckeditor/ckeditor5-code-block/src/codeblock.js';
+import Essentials from '@ckeditor/ckeditor5-essentials/src/essentials.js';
+import FindAndReplace from '@ckeditor/ckeditor5-find-and-replace/src/findandreplace.js';
+import Heading from '@ckeditor/ckeditor5-heading/src/heading.js';
+import HorizontalLine from '@ckeditor/ckeditor5-horizontal-line/src/horizontalline.js';
+import Image from '@ckeditor/ckeditor5-image/src/image.js';
+import ImageCaption from '@ckeditor/ckeditor5-image/src/imagecaption.js';
+import ImageResize from '@ckeditor/ckeditor5-image/src/imageresize.js';
+import ImageStyle from '@ckeditor/ckeditor5-image/src/imagestyle.js';
+import ImageToolbar from '@ckeditor/ckeditor5-image/src/imagetoolbar.js';
+import Indent from '@ckeditor/ckeditor5-indent/src/indent.js';
+import Italic from '@ckeditor/ckeditor5-basic-styles/src/italic.js';
+import Link from '@ckeditor/ckeditor5-link/src/link.js';
+import LinkImage from '@ckeditor/ckeditor5-link/src/linkimage.js';
+import List from '@ckeditor/ckeditor5-list/src/list.js';
+import ListStyle from '@ckeditor/ckeditor5-list/src/liststyle.js';
+import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph.js';
+import PasteFromOffice from '@ckeditor/ckeditor5-paste-from-office/src/pastefromoffice.js';
+import RemoveFormat from '@ckeditor/ckeditor5-remove-format/src/removeformat.js';
+import Strikethrough from '@ckeditor/ckeditor5-basic-styles/src/strikethrough.js';
+import Subscript from '@ckeditor/ckeditor5-basic-styles/src/subscript.js';
+import Superscript from '@ckeditor/ckeditor5-basic-styles/src/superscript.js';
+import Table from '@ckeditor/ckeditor5-table/src/table.js';
+import TableCaption from '@ckeditor/ckeditor5-table/src/tablecaption.js';
+import TableCellProperties from '@ckeditor/ckeditor5-table/src/tablecellproperties';
+import TableProperties from '@ckeditor/ckeditor5-table/src/tableproperties';
+import TableToolbar from '@ckeditor/ckeditor5-table/src/tabletoolbar.js';
+import TextTransformation from '@ckeditor/ckeditor5-typing/src/texttransformation.js';
+import TodoList from '@ckeditor/ckeditor5-list/src/todolist';
+import Underline from '@ckeditor/ckeditor5-basic-styles/src/underline.js';
+import WordCount from '@ckeditor/ckeditor5-word-count/src/wordcount.js';
+import {ImageInsert} from "@ckeditor/ckeditor5-image";
+
+class Editor extends ClassicEditor {
+}
+
+// Plugins to include in the build.
+Editor.builtinPlugins = [
+    Alignment,
+    Autoformat,
+    BlockQuote,
+    Bold,
+    CKFinderUploadAdapter,
+    Code,
+    CodeBlock,
+    Essentials,
+    FindAndReplace,
+    Heading,
+    HorizontalLine,
+    Image,
+    ImageCaption,
+    ImageResize,
+    ImageStyle,
+    ImageToolbar,
+    ImageInsert,
+    Indent,
+    Italic,
+    Link,
+    LinkImage,
+    List,
+    ListStyle,
+    Paragraph,
+    PasteFromOffice,
+    RemoveFormat,
+    Strikethrough,
+    Subscript,
+    Superscript,
+    Table,
+    TableCaption,
+    TableCellProperties,
+    TableProperties,
+    TableToolbar,
+    TextTransformation,
+    TodoList,
+    Underline,
+    WordCount
+];
+
+// Editor configuration.
+Editor.defaultConfig = {
+    toolbar: {
+        items: [
+            'heading',
+            '|',
+            'undo',
+            'redo',
+            '|',
+            'findAndReplace',
+            '|',
+            'bold',
+            'code',
+            'italic',
+            'underline',
+            'strikethrough',
+            'superscript',
+            'subscript',
+            'removeFormat',
+            '|',
+            'outdent',
+            'indent',
+            'alignment',
+            'bulletedList',
+            'numberedList',
+            'todoList',
+            '|',
+            'insertImage',
+            'blockQuote',
+            'insertTable',
+            'horizontalLine',
+            'link',
+            'codeBlock'
+        ]
+    },
+    language: 'zh-cn',
+    image: {
+        toolbar: [
+            'imageTextAlternative',
+            'imageStyle:inline',
+            {
+                name: 'blockDropdown',
+                title: 'blockDropdown',
+                items: ['imageStyle:alignBlockLeft', 'imageStyle:block', 'imageStyle:alignBlockRight'],
+                defaultItem: 'imageStyle:block'
+            },
+            {
+                name: 'sideDropdown',
+                title: 'sideDropdown',
+                items: ['imageStyle:alignLeft', 'imageStyle:alignRight'],
+                defaultItem: 'imageStyle:alignRight'
+            },
+            '|',
+            'linkImage'
+        ]
+    },
+    table: {
+        contentToolbar: [
+            'tableColumn',
+            'tableRow',
+            'mergeTableCells',
+            'tableCellProperties',
+            'tableProperties'
+        ]
+    }
+};
+
+// noinspection JSUnusedGlobalSymbols
+export default Editor;